@@ -5,7 +5,10 @@
   "main": "main.js",
   "scripts": {
     "start": "node src/main.js",
-    "dev": "nodemon src/main.js"
+    "dev": "nodemon src/main.js",
+    "web:dev": "vite src/client",
+    "web:build": "vite build src/client",
+    "web:preview": "vite preview src/client"
   },
   "repository": {
     "type": "git",
@@ -27,13 +30,24 @@
     "helmet": "^5.1.1",
     "http-status": "^1.5.2",
     "morgan": "^1.10.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.3.0",
+    "react-toastify": "^9.0.8",
     "youtube-dl-exec": "^2.1.4"
   },
   "devDependencies": {
-    "eslint": "^8.20.0",
+    "@types/react": "^18.0.17",
+    "@types/react-dom": "^18.0.6",
+    "@vitejs/plugin-react": "^2.0.0",
+    "autoprefixer": "^10.4.8",
+    "eslint": "^8.21.0",
     "eslint-config-standard": "^17.0.0",
     "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-n": "^15.2.4",
-    "eslint-plugin-promise": "^6.0.0"
+    "eslint-plugin-promise": "^6.0.0",
+    "postcss": "^8.4.16",
+    "tailwindcss": "^3.1.8",
+    "vite": "^3.0.5"
   }
 }
