@@ -1,3 +1,13 @@
-export default function FAQ () {
-  return <div>FAQ</div>
+import React from 'react'
+
+import Container from '../../components/UI/container'
+
+const Faq = () => {
+  return (
+    <Container>
+      <div>FAQ</div>
+    </Container>
+  )
 }
+
+export default Faq
