@@ -1,9 +1,10 @@
 /** @type {import('tailwindcss').Config} */
 module.exports = {
-  content: ['./src/**/*.{html,js,jsx}'],
+  content: ['./src/client/**/*.{html,js,jsx}'],
   theme: {
     extend: {},
     container: {
+      center: true,
       padding: {
         DEFAULT: '1rem',
         sm: '2rem',
